Send and receive multi-fragment messages (#18)
Add `Connection::send` and `Connection::recv`, which return results for sending and receiving a whole message, respectively. These futures also implement message fragmentation and reassembly for large messages. I've included a test verifying that fragmentation and reassembly work, at least when fragments are sent in-order. These futures also work on the raw bytes, which is the first part of implementing #17.